Rearranged some files and started working on egg_hunt
